chore: Add endpoint to add word to topic
This commit adds a new endpoint to the `TopicController` class that allows users to add a word to a topic. The endpoint is `POST /{topicId}/add-word` and accepts a `Word` object in the request body. The method `addWordToTopic` in the `TopicController` class calls the `addWordToTopic` method in the `TopicService` class to save the word to the topic. This feature enhances the functionality of the application by allowing users to add words to topics.