chore: Delete word functionality in WordTable component
This commit adds the deleteWord function to the WordTable component in order to delete a word from the database. The function sends a DELETE request to the API endpoint `/api/words/{wordId}/delete` and removes the word from the local storage to update the UI. This enhancement improves the functionality of the application by allowing users to delete words from the word table.