feat: Add @JsonManagedReference to Course and User models
This commit adds the @JsonManagedReference annotation to the Course and User models in order to resolve the circular reference issue when serializing the enrollments list. By using @JsonManagedReference, the serialization process will include the enrollments list when serializing the Course and User objects, but exclude it when serializing the Enrollment objects. This enhancement improves the JSON serialization of the models and prevents infinite recursion errors.